[server-main] feat: Server Main foundation
This commit:
Set up the http and websocket server main.

Creates a way for websocket to be assigned id based
on client. The id will then be used to identify
which user will be sent the notification
according to their deepfake analyse result.

Sets up api to receive result from server model.