[web-interface] feat: Call Management
This commit:
Add a new feature that record the incoming voice
on the other side every 5 seconds and send it
to the server for analyse to see if the voice
is deepfake or not.

Add a new feature that allows the user to switch
between the audio input types such as microphone
and audio file to test the case where the attacker
or the deepfake user, initiate the call.

Add a new feature that allows the user to answer
the call with an audio file to test the case where
the attacker or the deepfake user answer the call.

Add a new feature that force one person to hangup
if the other person hangup to resemble a phone
call.