[web-interface] feat: Admin & Client View
This commit:

Installs react-router-dom and setup for web
interface navigation.

Installs bootstrap react for styling the web
interface.

Create a dedicated Admin view for testing the
real time call deepfake analytic feature of the
application.

Create a dedicated Client view to be embeded
within the react native application for the
real time call deepfake analytic feature of the
application.