[web] ref: Admin, AdminCall & Test
This commit:

Refactors the code of AdminCall, bundle the
majority of state together and create specific
useEffect to track state and update more
accurately, making the code more readable.

Fix an error where an admin can not log in with
the same id that they just loggod out due to the
lack of update of the clientID state when logging
out.

Delete redundant code in Test.js, an empty prop
object.