[database] Modify database
This commit:

Add table Block_Log into the
database to enable the user to block a contact,
even if the blocked user is not added as a
contact.

Add table Report_Log into the
database to enable the user to report a another
user which is suspicious of using deepfake.

Modify the data generation and insertion process
to incorporate the changes.