[server] ref: Websocket ID attachment
This commit changes the way id is attached to the
websocket. Previously, if the id sent back is a
negative number, to differentiate between the
mobile and web interface, it will be converted to
its positive counterpart. Now it will not be
converted so the deepfake analyse result will be
sent to the web interface.